refactor(navigation): move desktop source to other dir
