+++ src/_shell/navigation/desktop/category-item.tsx
