+++ src/_shell/navigation/desktop/index.ts
@@ -1,0 +1,1 @@
+export * from './nav-menu';
