+++ src/_shell/navigation/mobile/accordion/call-to-action-accordion-item.tsx
@@ -1,7 +1,7 @@
 import type { QRL } from '@builder.io/qwik';
 import { component$ } from '@builder.io/qwik';
 import { AccordionHeader, AccordionItem } from '@qwik-ui/headless';
-import type { NavTopMenuItem } from '../../../nav-menu/nav-top-menu-item.type';
+import type { NavTopMenuItem } from '../../desktop/nav-top-menu-item.type';
 
 type CallToActionAccordionItemProps = {
   key?: string;
