+++ src/_shell/navigation/mobile/accordion/link-accordion-item.tsx
@@ -2,7 +2,7 @@ import type { QRL } from '@builder.io/qwik';
 import { component$ } from '@builder.io/qwik';
 import { Link } from '@builder.io/qwik-city';
 import { AccordionHeader, AccordionItem } from '@qwik-ui/headless';
-import type { NavTopMenuItem } from '../../../nav-menu/nav-top-menu-item.type';
+import type { NavTopMenuItem } from '../../desktop/nav-top-menu-item.type';
 
 type CollapsibleAccordionItemProps = {
   item: NavTopMenuItem;
